+++ Runtime/Model/BacktraceData.cs
@@ -68,6 +68,12 @@ namespace Backtrace.Unity.Model
         public string[] Classifier;
 
         /// <summary>
+        /// Source code information - right now we support source code only for BacktraceUnhandledException exceptions.
+        /// </summary>
+        [JsonProperty(PropertyName = "sourceCode ")]
+        internal BacktraceSourceCode SourceCode;
+
+        /// <summary>
         /// Get a path to report attachments
         /// </summary>
         [JsonIgnore]
@@ -132,7 +138,8 @@ namespace Backtrace.Unity.Model
                 {"classifiers", new JArray(Classifier)},
                 {"attributes", Attributes.ToJson()},
                 {"annotations", Annotation.ToJson()},
-                {"threads", ThreadData == null ? null : ThreadData.ToJson()}
+                {"threads", ThreadData == null ? null : ThreadData.ToJson()},
+                {"sourceCode", SourceCode == null ? null : SourceCode.ToJson()}
             };
             return json.ToString();
         }
@@ -157,7 +164,8 @@ namespace Backtrace.Unity.Model
                 Classifier = classfiers,
                 Annotation = Annotations.Deserialize(@object["annotations"]),
                 Attributes = BacktraceAttributes.Deserialize(@object["attributes"]),
-                ThreadData = ThreadData.DeserializeThreadInformation(@object["threads"])
+                ThreadData = ThreadData.DeserializeThreadInformation(@object["threads"]),
+                SourceCode = BacktraceSourceCode.Deserialize(@object["sourceCode"]),
             };
         }
 
@@ -166,9 +174,13 @@ namespace Backtrace.Unity.Model
         /// </summary>
         private void SetThreadInformations()
         {
-            ThreadData = new ThreadData(Report.DiagnosticStack);
+            var faultingThread = !(Report.Exception is BacktraceUnhandledException
+                && string.IsNullOrEmpty(Report.Exception.StackTrace));
+
+            ThreadData = new ThreadData(Report.DiagnosticStack, faultingThread);
             ThreadInformations = ThreadData.ThreadInformations;
             MainThread = ThreadData.MainThread;
+            SourceCode = Report.SourceCode;
         }
 
         /// <summary>
@@ -178,7 +190,7 @@ namespace Backtrace.Unity.Model
         private void SetAttributes(Dictionary<string, object> clientAttributes)
         {
             Attributes = new BacktraceAttributes(Report, clientAttributes);
-            Annotation = new Annotations(Attributes.ComplexAttributes);
+            Annotation = new Annotations();
         }
 
         /// <summary>
@@ -194,7 +206,7 @@ namespace Backtrace.Unity.Model
             LangVersion = "Mono";
 #endif
 
-            AgentVersion = "2.1.1";
+            AgentVersion = "2.1.6";
             Classifier = Report.ExceptionTypeReport ? new[] { Report.Classifier } : null;
         }
     }