+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Backtrace Unity Release Notes
 
+## Version 2.1.6
+
+- Handling special case for string reports fingerprint when stack trace is empty.
+
+## Version 2.1.5
+
+- Backtrace Unity plugin UI improvements - added tooltips, headers and collapsible menu for advanced options.
+- Changed Client-side deduplication menu,
+- `BacktraceClient` now allows you to choose what type of fingerprint Backtrace should generate for reports without stack trace. `Use normalized exception message` allows you to use a normalized exception message to generate fingerprint, instead of stack trace.
+- Added exception source code information to exception and message type of reports.
+
+## Version 2.1.4
+
+- `EnvironmentVariable` class now will handle correctly nullable key/values,
+- `BacktraceAttributes` handle correctly nullable values.
+
+## Version 2.1.3
+
+- `BacktraceUnhandledException` will generate environment stack trace if Unity stack trace is empty. BacktraceReport will still generate normalized fingerprint for unhandled exception without stack trace.
+- `BacktraceUnhandledException` will provide information from Unity Error logger in source code property, which should improve error analysis in web debugger.
+- `BacktraceAttributes` won't try to collect `Annotations` anymore.
+- `Annotations` won't use ComplexAttributes property anymore.
+
+## Version 2.1.2
+
+- `BacktraceReport` will generate report fingerprint for exceptions without stack trace.
+- Changed game object depth default property value.
+- Added Exception information to the Annotation object.
+
 ## Version 2.1.1
 
 - UPM modifications - fixed editor assembly definition,
@@ -51,7 +80,7 @@
 - `BacktraceDatabase` retry method now respect correctly `BacktraceDatabase` `retryInterval` property,
 - New `Backtrace Configuration` won't override existing `Backtrace Configuration` in configuration directory.
 - Backtrace-Unity plugin tests now won't override some files in Backtrace-Database unit tests,
-- Backtrace-Unity plugin now allows you to setup client side deduplication rules via `Fingerprint`. By using this field you can limit reporting of an error that occurs many times over a few frames.
+- Backtrace-Unity plugin now allows you to setup client-side deduplication rules via `Fingerprint`. By using this field you can limit reporting of an error that occurs many times over a few frames.
 - Backtrace report limit watcher feature now will validate limits before BacktraceReport creation.
 - `BacktraceClient` and `BacktraceDatabase` now expose `Reload` method. You can use this method do dynamically change `BacktraceClient`/`BacktraceDatabase` configurations.